remontee des sources a la racine et corrections atoi et strsplit

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: ajouanna <ajouanna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/14 11:56:17 by ajouanna          #+#    #+#             */
-/*   Updated: 2016/11/14 14:50:32 by ajouanna         ###   ########.fr       */
+/*   Updated: 2016/11/15 11:37:13 by ajouanna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 int	ft_atoi(const char *nptr)
 {
 	int i;
-	int res;
+	long int res;
 	int sign;
 
 	res = 0;
